+++ jetty-core/jetty-server/src/main/config/modules/delay-until-content.mod
@@ -13,7 +13,7 @@ server
 server
 
 [after]
-threadlimit
+thread-limit
 
 [xml]
 etc/jetty-delayed.xml
