+++ jetty-core/jetty-server/src/main/config/modules/threadlimit.mod
@@ -1,24 +1,5 @@
-
 [description]
-Applies ThreadLimitHandler to entire server, to limit the threads per IP address for DOS protection.
-
-[tags]
-server
+DEPRECATED - use the thread-limit module instead.
 
 [depend]
-server
-
-[xml]
-etc/jetty-threadlimit.xml
-
-[ini-template]
-## Select style of proxy forwarded header
-#jetty.threadlimit.forwardedHeader=X-Forwarded-For
-#jetty.threadlimit.forwardedHeader=Forwarded
-
-## Enabled by default?
-#jetty.threadlimit.enabled=true
-
-## Thread limit per remote IP
-#jetty.threadlimit.threadLimit=10
-
+thread-limit
