+++ jetty-core/jetty-server/src/main/java/org/eclipse/jetty/server/handler/ThreadLimitHandler.java
@@ -46,8 +46,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * <p>Handler to limit the threads per IP address for DOS protection</p>
- * <p>The ThreadLimitHandler applies a limit to the number of Threads
+ * <p>Handler to limit the number of concurrent threads per remote IP address, for DOS protection.</p>
+ * <p>ThreadLimitHandler applies a limit to the number of threads
  * that can be used simultaneously per remote IP address.</p>
  * <p>The handler makes a determination of the remote IP separately to
  * any that may be made by the {@link ForwardedRequestCustomizer} or similar:</p>
@@ -103,7 +103,7 @@ public class ThreadLimitHandler extends ConditionalHandler.Abstract
         LOG.info(String.format("ThreadLimitHandler enable=%b limit=%d", _enabled, _threadLimit));
     }
 
-    @ManagedAttribute("true if this handler is enabled")
+    @ManagedAttribute("Whether this handler is enabled")
     public boolean isEnabled()
     {
         return _enabled;
